Restructure component/issue folder
