+++ src/ggrc-client/js/components/issue/issue-roles-wrapper/issue-roles-wrapper.js
